🚸(front) put new message route under [mailboxId]
Previously the new message view was not a child of the `[mailboxId]` route.
So as the selectedMailbox is computed from the url param `mailboxId`
selectedMailbox was empty on this view so the first mailbox was selected.
This was weird for users with several mailboxes as if this one wants to write
a message on a secondary mailbox, the left panel was updated to display the
default mailbox. So by nesting new message view under `[mailboxId]` we ensure
that the context is retain when the user wants to write a new message.