✨(frontend) allow to forward a message
Update MessageEditor and add actions to forward a message. Furthermore
attachments of the forwarded message are automatically included in the new
message.

Then when a user is replying or forwarding a message we want the user sees
that a message will be embedded. So we create a readonly custom block
for message composer that renders a preview of the embedded message.